type fixes for cross-compiling (e.g. with RIOT)

+++ src/ccnl-core/src/ccnl-content.c
@@ -50,9 +50,9 @@ ccnl_content_new(struct ccnl_pkt_s **pkt)
     char s[CCNL_MAX_PREFIX_SIZE];
     (void) s;
 
-    DEBUGMSG_CORE(TRACE, "ccnl_content_new %p <%s [%d]>\n",
+    DEBUGMSG_CORE(TRACE, "ccnl_content_new %p <%s [%lu]>\n",
              (void*) *pkt, ccnl_prefix_to_str((*pkt)->pfx, s, CCNL_MAX_PREFIX_SIZE),
-             ((*pkt)->pfx->chunknum) ? *((*pkt)->pfx->chunknum) : 0);
+             ((*pkt)->pfx->chunknum) ? (long unsigned) *((*pkt)->pfx->chunknum) : (long unsigned) 0);
 
     c = (struct ccnl_content_s *) ccnl_calloc(1, sizeof(struct ccnl_content_s));
     if (!c)
